selection layout implement
